Refactor TabMenu component
- Moved tab button rendering to a separate TabMenu component
- Passed activeTab prop to TabMenu for state control
-  Updated TabButton to highlight the active tab dynamically
- Improved separation of concerns and code reusability
