--- src/components/CoreConceptsType.ts
@@ -1,7 +1,0 @@
-interface CoreConceptsType {
-    title: string;
-    description: string;
-    image: string;
-}
-
-export default CoreConceptsType;
