--- src/components/TabButton.tsx
@@ -1,11 +1,0 @@
-import TabButtonProps from './TabButtonProps';
-
-const TabButton = ({ children, onButtonClick }: TabButtonProps) => {
-  return (
-    <li>
-      <button onClick={onButtonClick}>{children}</button>
-    </li>
-  );
-};
-
-export default TabButton;
