+++ src/components/core-concepts/CoreConceptsType.ts
@@ -1,0 +1,5 @@
+export interface CoreConceptsType {
+  title: string;
+  description: string;
+  image: string;
+}
