+++ src/components/tabs/TabButton.tsx
@@ -1,0 +1,13 @@
+import { TabButtonProps } from './TabButtonProps';
+
+const TabButton = ({ children, onButtonClick }: TabButtonProps) => {
+  return (
+    <li>
+      <button onClick={onButtonClick}>
+        <span>{children}</span>
+      </button>
+    </li>
+  );
+};
+
+export default TabButton;
